requirements.txt
#beautifulsoup4 = "^4.10.0"
#requests = "^2.28.2"
#PyPDF2 = "^1.26.0"
#datetime
#os

python = "^3.11"
python-dotenv==1.0.0
pytz==2023.3
python-telegram-bot==20.4
beautifulsoup4==4.12.2
requests==2.31.0
#python = "^3.11"
#telegram.ext
